// carrot-demo/src/interface.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type LengthCm = number;
export type DiameterCm = number;
export type AgeMonths = number;
export type Lat = number;
export type Long = number;
export type TemperatureDegc = number;
export type HumidityPct = number;

export interface Carrot {
  length_cm: LengthCm;
  diameter_cm: DiameterCm;
  age_months: AgeMonths;
  location: Location;
  conditions: Conditions;
  [k: string]: unknown;
}
export interface Location {
  lat: Lat;
  long: Long;
  [k: string]: unknown;
}
export interface Conditions {
  temperature_degC: TemperatureDegc;
  humidity_pct: HumidityPct;
  [k: string]: unknown;
}